Update tests to async name space
