+++ opensearchpy/__init__.pyi
@@ -28,8 +28,8 @@ import sys
 from typing import Tuple
 
 from .client import OpenSearch as OpenSearch
+from .connection import AsyncHttpConnection as AsyncHttpConnection
 from .connection import Connection as Connection
-from .connection import RequestsAsyncHttpConnection as RequestsAsyncHttpConnection
 from .connection import RequestsHttpConnection as RequestsHttpConnection
 from .connection import Urllib3HttpConnection as Urllib3HttpConnection
 from .connection_pool import ConnectionPool as ConnectionPool
@@ -58,8 +58,8 @@ try:
     from ._async.client import AsyncOpenSearch as AsyncOpenSearch
     from ._async.http_aiohttp import AIOHttpConnection as AIOHttpConnection
     from ._async.transport import AsyncTransport as AsyncTransport
-    from .helpers import AWSV4SignerAuth as AWSV4SignerAuth
     from .helpers import AWSV4SignerAsyncAuth as AWSV4SignerAsyncAuth
+    from .helpers import AWSV4SignerAuth as AWSV4SignerAuth
 except (ImportError, SyntaxError):
     pass
